requirements_production.txt
flask>=2.3.3
flask-sqlalchemy>=3.0.5
pandas>=2.2.3
numpy>=1.26.2
werkzeug>=2.3.7
requests>=2.31.0
pymysql>=1.1.0
cryptography>=41.0.0